Redirect after off, so it doesn't retrigger
If the app is loaded on the phone in the browser,
then the browser might redo the GET request to
show the page if it has been evicted from memory.
That would retrigger another Away event, which is
not desired. Now I am redireting to a fresh page
which can safely be reloaded without any side effects